rag-pdf-assistant/requirements.txt
openai
python-dotenv
langchain
flotorch-core